feat: request admin right

+++ src/main/java/com/jfam/subarashii/services/SecretQuestionService.java
@@ -1,8 +1,13 @@
 package com.jfam.subarashii.services;
 
+import com.jfam.subarashii.MyRunner;
+import com.jfam.subarashii.entities.Role;
 import com.jfam.subarashii.entities.SecretQuestion;
+import com.jfam.subarashii.entities.User;
 import com.jfam.subarashii.repositories.SecretQuestionRepository;
 import com.jfam.subarashii.utils.Constantes;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,6 +19,8 @@ import java.util.Optional;
 @Service
 public class SecretQuestionService {
 
+    private static final Logger logger = LoggerFactory.getLogger(MyRunner.class);
+
     @Autowired
     private SecretQuestionRepository secretQuestionRepository;
 
@@ -23,9 +30,15 @@ public class SecretQuestionService {
     }
 
 
-    public SecretQuestion createSecretQuestion(SecretQuestion secretQuestion) throws ResponseStatusException {
-        // Save SecretQuestion
-        return secretQuestionRepository.save(secretQuestion);
+    public SecretQuestion createSecretQuestion(User currentUser, SecretQuestion secretQuestion) throws ResponseStatusException {
+        if (currentUser.getRole().equals(Role.ADMIN.toString())) {
+            // Save SecretQuestion
+            return secretQuestionRepository.save(secretQuestion);
+
+        } else {
+            //User doesn't admin -> throw exception
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, Constantes.ErrorMessage.EXCEPTION_USER_DOESNT_RIGHTS_ADMIN);
+        }
     }
 
 
@@ -43,37 +56,42 @@ public class SecretQuestionService {
     }
 
 
-    public List<SecretQuestion> getAllSecretQuestions() {
-        return secretQuestionRepository.findAll();
-    }
+    public SecretQuestion patchSecretQuestionById(User currentUser, SecretQuestion secretQuestion, Long idSecretQuestion) throws ResponseStatusException{
+        if (currentUser.getRole().equals(Role.ADMIN.toString())) {
+            Optional<SecretQuestion> secretQuestionOpt = secretQuestionRepository.findById(idSecretQuestion);
+            if(secretQuestionOpt.isPresent()) {
+                SecretQuestion secretQuestionTemp = secretQuestionOpt.get();
+                secretQuestionTemp.setIdSecretQuestion(idSecretQuestion);
 
-
-    public SecretQuestion patchSecretQuestionById(SecretQuestion secretQuestion, Long idSecretQuestion) throws ResponseStatusException{
-        Optional<SecretQuestion> secretQuestionOpt = secretQuestionRepository.findById(idSecretQuestion);
-        if(secretQuestionOpt.isPresent()) {
-            SecretQuestion secretQuestionTemp = secretQuestionOpt.get();
-            secretQuestionTemp.setIdSecretQuestion(idSecretQuestion);
-
-            if (secretQuestion.getQuestion() != null){
-                secretQuestionTemp.setQuestion(secretQuestion.getQuestion());
+                if (secretQuestion.getQuestion() != null){
+                    secretQuestionTemp.setQuestion(secretQuestion.getQuestion());
+                }
+                return secretQuestionRepository.save(secretQuestionTemp);
+            } else {
+                //SecretQuestion doesn't exist -> throw exception
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, Constantes.ErrorMessage.EXCEPTION_SECRET_QUESTION_DOESNT_EXISTS);
             }
-            return secretQuestionRepository.save(secretQuestionTemp);
         } else {
-            //SecretQuestion doesn't exist -> throw exception
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, Constantes.ErrorMessage.EXCEPTION_SECRET_QUESTION_DOESNT_EXISTS);
+            //User doesn't admin -> throw exception
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, Constantes.ErrorMessage.EXCEPTION_USER_DOESNT_RIGHTS_ADMIN);
         }
     }
 
-    public boolean deleteSecretQuestionById(Long idSecretQuestion) throws ResponseStatusException {
-        Optional<SecretQuestion> secretQuestionOpt = secretQuestionRepository.findById(idSecretQuestion);
-        if(secretQuestionOpt.isPresent()) {
-            //SecretQuestion exist -> delete it
-            SecretQuestion secretQuestion = secretQuestionOpt.get();
-            secretQuestionRepository.delete(secretQuestion);
-            return true;
+    public boolean deleteSecretQuestionById(User currentUser, Long idSecretQuestion) throws ResponseStatusException {
+        if (currentUser.getRole().equals(Role.ADMIN.toString())) {
+            Optional<SecretQuestion> secretQuestionOpt = secretQuestionRepository.findById(idSecretQuestion);
+            if(secretQuestionOpt.isPresent()) {
+                //SecretQuestion exist -> delete it
+                SecretQuestion secretQuestion = secretQuestionOpt.get();
+                secretQuestionRepository.delete(secretQuestion);
+                return true;
+            } else {
+                //SecretQuestion doesn't exist -> throw exception
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, Constantes.ErrorMessage.EXCEPTION_SECRET_QUESTION_DOESNT_EXISTS);
+            }
         } else {
-            //SecretQuestion doesn't exist -> throw exception
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, Constantes.ErrorMessage.EXCEPTION_SECRET_QUESTION_DOESNT_EXISTS);
+            //User doesn't admin -> throw exception
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, Constantes.ErrorMessage.EXCEPTION_USER_DOESNT_RIGHTS_ADMIN);
         }
     }
 }
